Proyecto/backend/sql-python/requirements.txt
requests==2.22.0
six==1.13.0
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3
Flask==2.0.1
mysql-connector-python==8.0.26
python-dotenv==0.19.0
Flask-Cors==3.0.10
boto3==1.19.5
